requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
python-dotenv>=1.0.0
streamlit>=1.30.0
pypdf>=3.17.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.1 